Move ErrorListItem to component.
